requirements.txt
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=11.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise>=6.6.0